+++ openwisp_monitoring/device/migrations/0003_update_template.py
@@ -1,5 +1,7 @@
 # Manually Created
 
+from collections import OrderedDict
+
 from django.db import migrations
 
 from . import (
@@ -17,13 +19,27 @@ def migrate_data(apps, schema_editor):
     Template = apps.get_model('config', 'Template')
     try:
         template = Template.objects.get(pk=TEMPLATE_MONITORING_UUID)
-        template.config['files'][0] = TEMPLATE_OPENWISP_MONITORING_02
-        template.config['files'][1] = TEMPLATE_NETJSON_MONITORING_02
-        template.config['files'][2] = TEMPLATE_CRONTAB_MONITORING_02
-        template.config['files'][3] = TEMPLATE_RC_LOCAL_02
-        template.config['files'][4] = TEMPLATE_UPDATE_OPENWISP_PACKAGES_02
-        template.config['files'][5] = TEMPLATE_POST_RELOAD_HOOK_02
-        template.config['openwisp'][0]['included_interfaces'] = '*'
+        template.config = OrderedDict(
+            {
+                'files': [
+                    TEMPLATE_OPENWISP_MONITORING_02,
+                    TEMPLATE_NETJSON_MONITORING_02,
+                    TEMPLATE_CRONTAB_MONITORING_02,
+                    TEMPLATE_RC_LOCAL_02,
+                    TEMPLATE_UPDATE_OPENWISP_PACKAGES_02,
+                    TEMPLATE_POST_RELOAD_HOOK_02,
+                ],
+                'openwisp': [
+                    OrderedDict(
+                        {
+                            'config_name': 'monitoring',
+                            'config_value': 'monitoring',
+                            'included_interfaces': '*',
+                        }
+                    )
+                ],
+            }
+        )
         template.full_clean()
         template.save()
     except Template.DoesNotExist: