+++ CHANGES.rst
@@ -15,6 +15,9 @@ Backward incompatible changes
 - *Monitoring Template* is deprecated in favour of `openwisp monitoring packages <https://github.com/openwisp/openwrt-openwisp-monitoring#openwrt-openwisp-monitoring>`_.
   Follow the migration guide in `Migrating from monitoring scripts to monitoring packages <#migrating-from-monitoring-scripts-to-monitoring-packages>`_
   section of openwisp-monitoring documentation.
+- If you have made changes to the default *Monitoring Template*, then
+  create a backup of your template before running migrations. Running
+  migrations will make changes to the default *Monitoring Template*.
 
 Version 0.1.0 [2021-01-31]
 --------------------------